dz8 fix1

+++ 2sem/dz8/task.cpp
@@ -1,15 +1,15 @@
 #include "students.h"
 #include <iostream>
 
-size_t findNewElders(Student** year1, size_t n1,
-                     Student** year2, size_t n2,
+int findNewElders(Student** year1, int n1,
+                     Student** year2, int n2,
                      Student** newElders) {
-    size_t k = 0;
-    for (size_t i = 0; i < n2; ++i) {
+    int k = 0;
+    for (int i = 0; i < n2; ++i) {
         if (year2[i]->isElder != 1) continue; // не староста во 2-м году
 
         bool wasElderInYear1 = false;
-        for (size_t j = 0; j < n1; ++j) {
+        for (int j = 0; j < n1; ++j) {
             if (year1[j]->name == year2[i]->name &&
                 year1[j]->group + 100 == year2[i]->group) {
                 if (year1[j]->isElder == 1) {
@@ -34,9 +34,9 @@ size_t findNewElders(Student** year1, size_t n1,
     return k;
 }
 
-void sort(Student **students, size_t n){
-   for(size_t i=0;i<n;i++) {
-        for(size_t j=0;j<n-1;j++) {
+void sort(Student **students, int n){
+   for(int i=0;i<n;i++) {
+        for(int j=0;j<n-1;j++) {
             if(students[j]->group>students[j+1]->group){
                 std::swap(students[j],students[j+1]);
             }
@@ -44,14 +44,14 @@ void sort(Student **students, size_t n){
     }
 }
 
-void free(Student** a, size_t n){
-    for(size_t i = 0; i < n; i++) {
+void free(Student** a, int n){
+    for(int i = 0; i < n; i++) {
         delete a[i];
     }
 }
 
-void print(Student** a, size_t n){
-    for(size_t i = 0; i < n; i++) {
+void print(Student** a, int n){
+    for(int i = 0; i < n; i++) {
         std::cout<<*a[i]<<"\n";
     }
     puts("_______________");
